Refactor GoodbyeRules
